backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Marshmallow==0.15.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
PyJWT==2.11.0
bcrypt==4.1.2
Werkzeug==3.0.1
